requirements.txt
flask==2.3.3
python-dotenv==1.0.0
pillow==10.0.0
matplotlib==3.8.0
azure-ai-vision-imageanalysis==1.0.0b1
